+++ docs/changelog.md
@@ -1,5 +1,9 @@
 ## CHANGELOG
 
+### UNRELEASED
+
+ - Fix Service.get_one
+
 ### 0.16.0 - 2020-06-25
 
  - Create Service.exists_many method